+++ src/io/tanners/tax/TaxData.java
@@ -2,6 +2,8 @@ package io.tanners.tax;
 
 import io.tanners.tax.exception.ValueIsNegativeException;
 
+import java.math.BigDecimal;
+
 /**
  * Data class to hold the information about an item with it's tax information for later use
  */
@@ -9,13 +11,13 @@ public abstract class TaxData
 {
     protected int mQuantity;
     protected String mItem;
-    protected double mPrice;
+    protected BigDecimal mPrice;
 
     public abstract int getmQuantity();
     public abstract void setmQuantity(int mQuantity) throws ValueIsNegativeException;
     public abstract String getmItem();
     public abstract void setmItem(String mItem);
-    public abstract double getmPrice();
-    public abstract void setmPrice(double mPrice) throws ValueIsNegativeException;
+    public abstract BigDecimal getmPrice();
+    public abstract void setmPrice(BigDecimal mPrice) throws ValueIsNegativeException;
 
 }
